Moving from package 'api' to package 'resources/apimodel'
