+++ admin-backend/src/main/java/se/cag/labs/cagrms/admin/resources/UserResource.java
@@ -1,7 +1,7 @@
 package se.cag.labs.cagrms.admin.resources;
 
 
-import se.cag.labs.cagrms.admin.api.apimodel.User;
+import se.cag.labs.cagrms.admin.resources.apimodel.User;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
